Move AttrSeries from reporting.utils to (new) reporting.attrseries
